// src/types/pokemon.ts
export interface PokemonListResponse {
    count: number;
    next: string |null;
    previous: string | null;
    results: PokemonListItem[];
}


export interface PokemonListItem{
    name: string;
    url: string;
}

export interface PokemonType{
    slot: number;
    type:{
        name:string;
        url:string;
    };
}

export interface PokemonStat{
    base_stat: number;
    stat:{
        name:string;
        url:string;
    };
}

export interface OfficialArtwork {
  front_default: string;
}


export interface PokemonSprites{
    front_default: string;
    other?: {
        'official-artwork'?:{
            front_default: string;
        };
    };
}

export interface PokemonDetail{
    id: number;
    name: string;
    height: number;
    weight: number;
    base_experience: number;
    types: PokemonType[];
    sprites: PokemonSprites;
    stats: PokemonStat[];
}